feat: implement API error handling and storage services
