+++ roamance/src/constants/keys.ts
@@ -1,0 +1,5 @@
+export const STORAGE_KEYS = {
+  TOKEN: 'token',
+  USER: 'user',
+  TOKEN_EXPIRY: 'tokenExpiry',
+} as const;
